Predict the year songs are released
Music is a business, and it is a very lucrative one. Tens of thousands of songs are released every day, meaning there is significant competition to be heard and to draw listeners to specific songs. A main reason Spotify is so successful in the streaming age is due to their great playlists. The goal of my project is to utilize many sonic aspects of music tracks to make predictions about genre, popularity, and when the song was released. This will help connect newly added songs to the most beneficial playlists. The benefits here are twofold. First, it helps businesses curate musically consistent playlists, with the goal of users listening on the platform longer. This allows them to have a competitive edge over other music streaming services, as well as providing them with the ability to predict which songs would do well in specific playlists. Second, it will help songwriters understand what qualities they can implement into their compositions to cater to certain groups of listeners or to get on certain playlists.